requirements.txt
# $ pip install -r requirements.txt

digi-xbee~=1.5.0
sqlalchemy~=2.0.34
